requirements.txt
matplotlib==3.5.1
numba==0.60.0
numpy==2.2.5
opencv_python==4.9.0.80
pandas==1.2.4
Pillow==11.2.1
scikit_image==0.19.3
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
